requirements.txt
fastapi
uvicorn
tensorflow==2.16.1
scikit-learn==1.6.1
joblib
numpy
Pillow
python-multipart
